sigenie/pyproject.toml
[tool.poetry]
name = "sigenie"
version = "0.1.0"
description = ""
authors = ["Kwang Yong Jung <gyjong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
pydantic = "^2.9.2"
pypdf = "^4.3.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-groq = "^0.2.0"
langchain-anthropic = "^0.2.1"
langchain-google-genai = "^2.0.0"
langchain-cerebras = "^0.2.0"
langchain-experimental = "^0.3.0"
langchain_mongodb = "^0.2.0"
langgraph = "^0.2.22"
google-generativeai = "^0.7.1"
tabulate = "^0.9.0"
faiss-cpu = "^1.8.0.post1"
pymongo = "^4.9.1"
bs4 = "^0.0.2"
matplotlib = "^3.9.1"
pyvis = "^0.3.2"
networkx = "^3.3"
graphviz = "^0.20.3"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
streamlit = "^1.36.0"
# duckduckpy = "^0.2"
# duckduckgo-search = "^6.2.1"
# speechrecognition = "^3.10.4"
# gtts = "^2.5.2"
# pygame = "^2.6.0"
# pyttsx3 = "^2.90"
#neo4j = "^5.23.1"
pymupdf = "^1.24.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
